requirements.txt
Flask
python-dotenv
deep-translator
google-generativeai
